+++ uniflow/flow/transform/__init__.py
@@ -3,4 +3,4 @@
 # this register all possible model server into ModelServerFactory through
 # ModelServerFactory.register(cls.__name__, cls) in AbsModelServer
 # __init_subclass__
-from uniflow.model.server import *  # noqa: F401, F403
+from uniflow.flow.transform.utils.model_server import *  # noqa: F401, F403
